Introduce a new sort of assertion, testasserts, which makes it a bit easier to capture potentially-useful information in a test assertion without needing extra logging,
This is basically a kind of lazy expression capture thing, allowing us
to do some actual metaprogramming for a change. The only moderately
clever part is making sure that x&&y does the right kind of short
circuiting, which is usually a bit tricky with C++ operator
overloading.

I've only done overloads for a couple of operators here. I figure we
can add the later when they're actually needed; it's not very hard.