Deliberately make the predictability test less predictable.
Or at least, increase the SD, because that reduces the number of very
large outliers. This increases SD from 2.4ms to 3.1 but decreases the
excess kurtosis from 12.8 to -0.2. Also a useful reduction in the
median->99% gap. That makes the test pass much more reliably. Current
theory is that it's mostly down to reducing thundering herd effects.

There is some risk here that I'm just washing out a real effect rather
than fixing it. More investigation to come.